Update README.md
Hi, I noticed that the `Install with Sketchpacks` link was not working.  It looks like it was due to a name change in this repository.
I also noticed that the readme images are not being rendered on the plugins page on Sketchpacks.  This is due to relative URLs in the readme.

Changed:
- Updated `Install with Sketchpacks` link to use the `SketchContentSync.sketchplugin` name instead of `SketchContentSync`.
- Updated images to use full URLs instead of relative URLs. This will allow Sketchpacks to render the images correctly @ https://sketchpacks.com/contentsync/SketchContentSync.sketchplugin